scripts/export-import/requirements.txt
certifi
python-dotenv
rich
edgedb
